Wait for the databases instead of creating them when a host is set
When the database host is not the local host the databases can not be
set up; it is only possible to wait for them to be up.

Waiting for the database is done calling the "wait-for-connection"
script from the External files app; it is the same technique used in the
"autotest.sh" script from the server.

This change affects only MySQL and PostgreSQL databases; Oracle was not
taken into account.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>